+++ src/main/java/liadov/mypackage/Main.java
@@ -1,4 +1,4 @@
-package liadov.my_package;
+package liadov.mypackage;
 
 public class Main {
     public static void main(String[] args) {
